User story 2 complete, added poro for holidays and a new service ruby model for nader holidays. On user story 3
