+++ app/services/git_service.rb
@@ -1,9 +1,5 @@
 class GitService
-
-  def self.upcoming_holidays
-    input = get_data('https://date.nager.at/Api/v2/NextPublicHolidays/US')[0..2]
-    Holiday.upcoming(input)
-  end
+  extend GetUrl
 
   def self.repo_name
     get_data('https://api.github.com/repos/duguy8/little-esty-shop')[:name]
@@ -24,13 +20,5 @@ class GitService
 
   def self.prs
     get_data("https://api.github.com/repos/duguy8/little-esty-shop/pulls?state=all").length
-  end
-
-  def self.get_data(url)
-    response = Faraday.get(url) do |req|
-      req.headers['Authorization'] = "token #{ENV['github_token']}"
-    end
-    data = response.body
-    JSON.parse(data, symbolize_names: true)
   end
 end
