+++ app/views/merchants/discounts/_discounts.html.erb
@@ -1,7 +1,7 @@
 <ul>
   <% merchant.discounts.each do |discount| %>
     <li><%= link_to "#{discount.name}", merchant_discount_path(merchant, discount) %>
-    <%= discount.quantity_threshold %>
-    <%= discount.percentage_discount %></li>
+    Quantity Threshold: <%= discount.quantity_threshold %>,
+    Percentage Discount: <%= discount.percentage_discount %></li>
   <% end %>
 </ul>
