+++ spec/features/discounts/api/upcoming_holidays_spec.rb
@@ -12,11 +12,12 @@ RSpec.describe "When I visit the discounts index page" do
   describe "I see a section ~Upcoming Holidays~ with next 3 upcoming US holidays listed" do
     it "Is from the Next public holidays endpoint in Nager.Date API" do
       visit merchant_discounts_path(@merchant)
+      save_and_open_page
 
       within(".upcoming_holidays") do
-        expect(page).to have_content("Upcoming Holiday 1:")
-        expect(page).to have_content("Upcoming Holiday 2:")
-        expect(page).to have_content("Upcoming Holiday 3:")
+        expect(page).to have_content("#{NagerService.upcoming_holidays[0].name} on: #{NagerService.upcoming_holidays[0].date}")
+        expect(page).to have_content("#{NagerService.upcoming_holidays[1].name} on: #{NagerService.upcoming_holidays[1].date}")
+        expect(page).to have_content("#{NagerService.upcoming_holidays[2].name} on: #{NagerService.upcoming_holidays[2].date}")
       end
     end
   end
